refactor: remove unused imports and consolidate statistic functions
